requirements.txt
instagrapi
groq
python-dotenv
yt-dlp
python-vlc
